requirements.txt
python-dotenv
voyageai
aiofiles
chromadb==0.5.3
